Fixed minor 64-bit warnings.
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@9573 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/SharedFolder.cpp
@@ -162,7 +162,7 @@ CLibraryFolder* CLibraryFolder::GetFolderByName(LPCTSTR pszName) const
 	LPCTSTR szNextName = _tcschr( pszName, _T( '\\' ) );
 	if ( szNextName )
 	{
-		CString strName( pszName, szNextName - pszName );
+		CString strName( pszName, (int)( szNextName - pszName ) );
 		if ( m_pFolders.Lookup( strName, pOutput ) )
 			return pOutput->GetFolderByName( szNextName + 1 );
 	}
